+++ openhands_aci/navigator/__init__.py
@@ -1,0 +1,3 @@
+from .navigator import SymbolNavigator
+
+_GLOBAL_NAVIGATOR = SymbolNavigator()
