+++ openhands_aci/utils/shell.py
@@ -2,8 +2,8 @@ import os
 import subprocess
 import time
 
+from openhands_aci.core.results import maybe_truncate
 from openhands_aci.editor.config import MAX_RESPONSE_LEN_CHAR
-from openhands_aci.editor.results import maybe_truncate
 
 
 def run_shell_cmd(
